shadow-dom/focus/focus-method-delegatesFocus.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focus on shadow host with delegatesFocus</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>
<div id="host">
  <div id="slottedToSecondSlot" slot="secondSlot">slottedToSecondSlot</div>
  <div id="slottedToFirstSlot" slot="firstSlot">slottedToFirstSlot</div>
</div>
<div id="outside">outside</div>
</body>

<script>
const host = document.getElementById("host");
const slottedToSecondSlot = document.getElementById("slottedToSecondSlot");
const slottedToFirstSlot = document.getElementById("slottedToFirstSlot");
const outside = document.getElementById("outside");

const shadowRoot = host.attachShadow({ mode: "open", delegatesFocus: true });
const aboveSlots = document.createElement("div");
aboveSlots.innerText = "aboveSlots";
const firstSlot = document.createElement("slot");
firstSlot.name = "firstSlot";
const secondSlot = document.createElement("slot");
secondSlot.name = "secondSlot";
const belowSlots = document.createElement("div");
belowSlots.innerText = "belowSlots";
shadowRoot.appendChild(aboveSlots);
shadowRoot.appendChild(firstSlot);
shadowRoot.appendChild(secondSlot);
shadowRoot.appendChild(belowSlots);

const elementsInFlatTreeOrder = [host, aboveSlots, firstSlot,
  slottedToFirstSlot, secondSlot, slottedToSecondSlot, belowSlots, outside];

// Final structure:
// <div #host> (delegatesFocus=true)
//    #shadowRoot
//      <div #aboveSlots>
//      <slot #firstSlot>
//        (slotted) <div #slottedToFirstSlot>
//      <slot #secondSlot>
//        (slotted) <div #slottedToSecondSlot>
//      <div #belowSlots>
// <div #outside>

function setTabIndex(elements, value) {
  for (const el of elements) {
    el.tabIndex = value;
  }
}

function setAllTabIndex(value) {
  setTabIndex(elementsInFlatTreeOrder, value);
}

function removeAllTabIndex() {
  for (const el of elementsInFlatTreeOrder) {
    el.removeAttribute("tabindex");
  }
}

function resetFocus() {
  removeAllTabIndex();
  if (shadowRoot.activeElement)
    shadowRoot.activeElement.blur();
  if (document.activeElement)
    document.activeElement.blur();
}

test(() => {
  resetFocus();
  setAllTabIndex(0);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots> tabindex=0
  //      <slot #firstSlot> tabindex=0
  //        (slotted) <div #slottedToFirstSlot> tabindex=0
  //      <slot #secondSlot> tabindex=0
  //        (slotted) <div #slottedToSecondSlot> tabindex=0
  //      <div #belowSlots> tabindex=0
  // <div #outside> tabindex=0
  // flattened tabindex-ordered focus navigation scope of #host:
  // [aboveSlots, slottedToFirstSlot, slottedToSecondSlot, belowSlots]
  host.focus();
  assert_equals(shadowRoot.activeElement, aboveSlots);
  assert_equals(document.activeElement, host);
}, "focus() on host with delegatesFocus, all tabindex=0");

test(() => {
  resetFocus();
  setAllTabIndex(0);
  setTabIndex([host], -1);
  // flattened tabindex-ordered focus navigation scope of #host:
  // [aboveSlots, slottedToFirstSlot, slottedToSecondSlot, belowSlots]
  host.focus();
  assert_equals(shadowRoot.activeElement, aboveSlots);
  assert_equals(document.activeElement, host);
}, "focus() on host with delegatesFocus & tabindex =-1, all other tabindex=0");

test(() => {
  resetFocus();
  setTabIndex([aboveSlots, slottedToFirstSlot, slottedToSecondSlot, belowSlots], 0);
  // flattened tabindex-ordered focus navigation scope of #host:
  // [aboveSlots, slottedToFirstSlot, slottedToSecondSlot, belowSlots]
  host.focus();
  assert_equals(shadowRoot.activeElement, aboveSlots);
  assert_equals(document.activeElement, host);
}, "focus() on host with delegatesFocus & no tabindex, all other tabindex=0");

test(() => {
  resetFocus();
  setAllTabIndex(-1);
  setTabIndex([host], 0);
  // flattened tabindex-ordered focus navigation scope of #host: []
  host.focus();
  assert_equals(shadowRoot.activeElement, null);
  assert_equals(document.activeElement, null);
}, "focus() on host with delegatesFocus & tabindex = 0, all other tabindex=-1");

test(() => {
  resetFocus();
  removeAllTabIndex();
  // flattened tabindex-ordered focus navigation scope of #host: []
  host.focus();
  assert_equals(shadowRoot.activeElement, null);
  assert_equals(document.activeElement, null);
}, "focus() on host with delegatesFocus, all without tabindex");

test(() => {
  resetFocus();
  // flattened tabindex-ordered focus navigation scope of #host: []
  setAllTabIndex(-1);
  host.focus();
  assert_equals(shadowRoot.activeElement, null);
  assert_equals(document.activeElement, null);
}, "focus() on host with delegatesFocus, all tabindex=-1");

test(() => {
  resetFocus();
  setAllTabIndex(-1);
  setTabIndex([host, belowSlots], 0);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots>
  //      <slot #firstSlot>
  //        (slotted) <div #slottedToFirstSlot>
  //      <slot #secondSlot>
  //        (slotted) <div #slottedToSecondSlot>
  //      <div #belowSlots> tabindex=0
  // <div #outside>
  // flattened tabindex-ordered focus navigation scope of #host:
  // [belowSlots]
  host.focus();
  assert_equals(shadowRoot.activeElement, belowSlots);
  assert_equals(document.activeElement, host);
}, "focus() on host with delegatesFocus & tabindex=0, #belowSlots with tabindex=0");

test(() => {
  resetFocus();
  setAllTabIndex(-1);
  setTabIndex([host, outside], 0);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots>
  //      <slot #firstSlot>
  //        (slotted) <div #slottedToFirstSlot>
  //      <slot #secondSlot>
  //        (slotted) <div #slottedToSecondSlot>
  //      <div #belowSlots>
  // <div #outside> tabindex=0
  // flattened tabindex-ordered focus navigation scope of #host:
  // []
  host.focus();
  assert_equals(shadowRoot.activeElement, null);
  assert_equals(document.activeElement, null);
}, "focus() on host with delegatesFocus & tabindex=0, #outside with tabindex=0");

test(() => {
  resetFocus();
  setTabIndex([host, aboveSlots, belowSlots], 0);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots> tabindex=0
  //      <slot #firstSlot>
  //        (slotted) <div #slottedToFirstSlot>
  //      <slot #secondSlot>
  //        (slotted) <div #slottedToSecondSlot>
  //      <div #belowSlots> tabindex=0
  // <div #outside>
  // flattened tabindex-ordered focus navigation scope of #host:
  // [aboveSlots, belowSlots]
  host.focus();
  assert_equals(shadowRoot.activeElement, aboveSlots);
  assert_equals(document.activeElement, host);
}, "focus() on host with delegatesFocus & tabindex=0, #aboveSlots and #belowSlots with tabindex=0");

test(() => {
  resetFocus();
  setAllTabIndex(-1);
  setTabIndex([host, aboveSlots], 0);
  setTabIndex([belowSlots], 1);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots> tabindex=0
  //      <slot #firstSlot>
  //        (slotted) <div #slottedToFirstSlot>
  //      <slot #secondSlot>
  //        (slotted) <div #slottedToSecondSlot>
  //      <div #belowSlots> tabindex=1
  // <div #outside>
  // flattened tabindex-ordered focus navigation scope of #host:
  // [belowSlots, aboveSlots]
  host.focus();
  assert_equals(shadowRoot.activeElement, belowSlots);
  assert_equals(document.activeElement, host);
}, "focus() on host with delegatesFocus & tabindex=0, #aboveSlots with tabindex=0, #belowSlots with tabindex=1");

test(() => {
  resetFocus();
  setTabIndex([host, firstSlot, slottedToFirstSlot, secondSlot, slottedToSecondSlot], 0);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots>
  //      <slot #firstSlot> tabindex=0
  //        (slotted) <div #slottedToFirstSlot> tabindex=0
  //      <slot #secondSlot> tabindex=0
  //        (slotted) <div #slottedToSecondSlot> tabindex=0
  //      <div #belowSlots>
  // <div #outside>
  // flattened tabindex-ordered focus navigation scope of #host:
  // [slottedToFirstSlot, slottedToSecondSlot]
  host.focus();
  assert_equals(shadowRoot.activeElement, null);
  assert_equals(document.activeElement, slottedToFirstSlot);
}, "focus() on host with delegatesFocus & tabindex=0, #firstSlot, #slottedToFirstSlot, #secondSlot, #slottedToSecondSlot with tabindex=0");

test(() => {
  resetFocus();
  setTabIndex([firstSlot], 1);
  setTabIndex([host, aboveSlots, slottedToFirstSlot], 0);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots> tabindex=0
  //      <slot #firstSlot> tabindex=1
  //        (slotted) <div #slottedToFirstSlot> tabindex=0
  //      <slot #secondSlot>
  //        (slotted) <div #slottedToSecondSlot>
  //      <div #belowSlots>
  // <div #outside>
  // flattened tabindex-ordered focus navigation scope of #host:
  // [slottedToFirstSlot, aboveSlots]
  host.focus();
  assert_equals(shadowRoot.activeElement, null);
  assert_equals(document.activeElement, slottedToFirstSlot);
}, "focus() on host with delegatesFocus & tabindex=0, #firstSlot with tabindex=1, #slottedToFirstSlot and #aboveSlots with tabindex=0");

test(() => {
  resetFocus();
  setTabIndex([slottedToFirstSlot], 1);
  setTabIndex([host, aboveSlots], 0);
  // Structure:
  // <div #host> (delegatesFocus=true) tabindex=0
  //    #shadowRoot
  //      <div #aboveSlots> tabindex=0
  //      <slot #firstSlot>
  //        (slotted) <div #slottedToFirstSlot> tabindex=1
  //      <slot #secondSlot>
  //        (slotted) <div #slottedToSecondSlot>
  //      <div #belowSlots>
  // <div #outside>
  // flattened tabindex-ordered focus navigation scope of #host:
  // [aboveSlots, slottedToFirstSlot]
  host.focus();
  assert_equals(shadowRoot.activeElement, aboveSlots);
  assert_equals(document.activeElement, host);
}, "focus() on host with delegatesFocus & tabindex=0, #slottedTofirstSlot with tabindex=1, #aboveSlots with tabindex=0");
</script>
